pyproject.toml
[project]
name = "booklore-enrich"
version = "0.1.0"
description = "Enrich BookLore library with metadata from romance.io and thebooknaut.com"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.1",
    "rich>=14.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "tomli-w>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.0",
]
scraping = [
    "playwright>=1.50.0",
]

[project.scripts]
booklore-enrich = "booklore_enrich.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/booklore_enrich"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
